Change in prepare.sh - Conditinal git commands

+++ shell/01_prepare.sh
@@ -3,10 +3,13 @@
 export PARTNER=$1
 export STAGE=$2
 
-git push origin --delete ${PARTNER}-${STAGE} &>/dev/null
-git branch -D ${PARTNER}-${STAGE} &>/dev/null
-git checkout -b ${PARTNER}-${STAGE}
-git push -u origin ${PARTNER}-${STAGE}
+if ! [ -z ${CODEBUILD_SOURCE_REPO_URL+x} ]
+then 
+    git push origin --delete ${PARTNER}-${STAGE} &>/dev/null
+    git branch -D ${PARTNER}-${STAGE} &>/dev/null
+    git checkout -b ${PARTNER}-${STAGE}
+    git push -u origin ${PARTNER}-${STAGE}
+fi
 
 export ENVIRONMENT=$1-$2
 export SDK_URL="https://$PARTNER.sdk.$STAGE.com/index.js"
